+++ contrib/gdm2s3m/src/s3m.h
@@ -80,7 +80,7 @@ struct S3M_sample {
 };
 
 /* function prototypes */
-uint8_t *save_s3m (struct S3M_file *s3m, uint32_t *stream_len);
+uint8_t *save_s3m (struct S3M_file *s3m, size_t *stream_len);
 struct S3M_file *convert_gdm_to_s3m (struct GDM_file *gdm);
 void free_s3m (struct S3M_file *s3m);
