+++ contrib/gdm2s3m/src/utility.c
@@ -41,7 +41,7 @@ static void alloc_to_stream (void *src, uint8_t **dest, uint32_t n)
   *dest += n;
 }
 
-void check_s_to_a (uint8_t *start, uint32_t size, void *dest, uint8_t **src,
+void check_s_to_a (uint8_t *start, size_t size, void *dest, uint8_t **src,
                    uint32_t n)
 {
   const size_t diff = *src - start;
@@ -54,7 +54,7 @@ void check_s_to_a (uint8_t *start, uint32_t size, void *dest, uint8_t **src,
   stream_to_alloc (dest, src, n);
 }
 
-void check_a_to_s (uint8_t **start, uint32_t *size, void *src, uint8_t **dest, uint32_t n)
+void check_a_to_s (uint8_t **start, size_t *size, void *src, uint8_t **dest, uint32_t n)
 {
   const size_t diff = *dest - *start;
