+++ src/render_gl1.c
@@ -186,7 +186,7 @@ static bool gl1_set_video_mode(struct graphics_data *graphics,
     if(!initialized)
     {
       const char *version, *extensions;
-      float version_float;
+      double version_float;
 
       version = (const char *)gl1.glGetString(GL_VERSION);
       if(!version)
